Refactoring and error handling added
